refactor(admin-ui): extract FileType and Placeholder into their own files
Completes the split started with Thumbnail. Moving FileType also takes the
ten asset imports with it, which were most of RichItemThumbnail's header;
what is left is the props type and the choice of which of the three to
render.
